catching yet correct links

+++ bin/setup.sh
@@ -10,13 +10,17 @@ vGITDIR=~/git
 vLIQPROM=false
 vLIQPROMLINK="https://github.com/nojhan/liquidprompt.git"
 
-# link dotfiles 
+# link dotfiles
 vDOTFILES=$vBASEDIR/dotfiles/.[!.]*
 for i in $vDOTFILES;do
     vORIGFILE=~/$(basename $i)
-    [ -e $vORIGFILE ] && mv $vORIGFILE $vORIGFILE.old \
-        && echo "saving existing file $vORIGFILE to $vORIGFILE.old ..."
-ln -s $i ~ && echo "creating symlink from $i to ~/$(basename $i) ..."
+    if ! [ -h $vORIGFILE -a $(readlink -f $vORIGFILE) = $i ];then
+    {
+        [ -e $vORIGFILE ] && mv $vORIGFILE $vORIGFILE.old \
+            && echo "saving existing file $vORIGFILE to $vORIGFILE.old ..."
+        ln -s $i ~ && echo "creating symlink from $i to ~/$(basename $i) ..."
+    }
+    else echo "$vORIGFILE: nothing to do ..."
 done
 
 # make bin folder and link binaries
@@ -40,7 +44,7 @@ done
        git clone -q $vLIQPROMLINK $vGITDIR && echo "cloning liquidprompt repository ..."
     fi
     ln -s $vBASEDIR/liquidprompt/.liquidpromptrc $vLIQPROMORIG \
-        && echo "creating symlink from $vBASEDIR/liquidprompt/.liquidpromptrc to $vLIQPROMORIG ..." 
+        && echo "creating symlink from $vBASEDIR/liquidprompt/.liquidpromptrc to $vLIQPROMORIG ..."
 }
 
 # finally source .bashrc
